move op-e2e/e2eutils/wait.go into op-service
