+++ op-service/client/utils/waits.go
@@ -1,4 +1,4 @@
-package e2eutils
+package utils
 
 import (
 	"context"
